requirements.txt
# API & server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0

# Binance
requests>=2.31.0

# DB
sqlalchemy>=2.0.25
pymysql>=1.1.0
cryptography>=41.0.0

# Telegram (step 8)
httpx>=0.26.0

# Optional: Redis for queue (MVP can use DB)
# redis>=5.0.0
